# Settable Defaults & A New Print Method #
### Setting Defaults ###

This update to version 2.10 / 2.11 adds the ability for to read the
default maximums from an XML file. That is, whenever the maximum
quantity for a given print type needs to be changed, I usually had to go
and hard-code the new defaults in the client program and then update all
PCs with the new client. Granted, this only happens once a year on
average, but it's still better to *not* have variables hard-coded
whenever possible.

Thus, I have learned how to have a Java program read from and write to
XML files. I chose XML since the data set needed for this (for now) is
very small and a new table in our database seems like overkill. It also
allows for a fallback of editing the XML file by hand if necessary,
rather than having to log into the SQL Server Management Studio and
typing out complicated SQL commands.

### Digital Cups ###

With the purchase of two "Helix" machines for printing images digitally
directly onto cups, we need a new print method. This has now been added
to the Art Dept program throughout.

### To Do ###

Shortly, the database table "Day" will be updated so four of its columns
will be renamed to help with disambiguation. This will necessitate
another update to the program since the SQL commands are embedded in the
classes.

Signed-off-by: Chris McGee <chris@chrismcgee.info>